🧹 Extract Drop class from useMatrixRain.ts to Drop.ts
